frer: add isdx (iflow) alloc/free subcommands
This depends on the appropriate change in the netlink API. ISDX is a
shared resource and if the user simply picks them at will, it causes
problems for other features which try to use the same ISDXs. Not to
mention the counters will never be cleared after use etc.

The member stream index is already explicitly managed by the frer tool,
so I thought this was the simplest change to hook into lan9645x isdx
allocation.